+++ src/Laul.Application/Services/LikeDislikes/Commands/UpdateLikeDislike/UpdateLikeDislikeCommand.cs
@@ -5,6 +5,6 @@ namespace Laul.Application.Services.LikeDislikes.Commands.UpdateLikeDislike
     public class UpdateLikeDislikeCommand : IRequest<Unit>
     {
         public Guid UserId { get; set; }
-        public int SongId { get; set; }
+        public ulong SongId { get; set; }
     }
 }
