+++ src/Laul.Application/Services/ListeningStats/Command/CreateListeningStat/CreateListeningStatCommand.cs
@@ -2,7 +2,7 @@
 
 namespace Laul.Application.Services.ListeningStats.Command.CreateListeningStat
 {
-    public class CreateListeningStatCommand : IRequest<int>
+    public class CreateListeningStatCommand : IRequest<ulong>
     {
         public ulong SongId { get; set; }
         public Guid UserId { get; set; }
