+++ src/Laul.Application/Services/Playlists/Commands/AddPlaylistSong/AddPlaylistSongCommand.cs
@@ -2,7 +2,7 @@
 
 namespace Laul.Application.Services.Playlists.Commands.AddPlaylistSong
 {
-    public class AddPlaylistSongCommand : IRequest<int>
+    public class AddPlaylistSongCommand : IRequest<ulong>
     {
         public Guid UserId { get; set; }
         public ulong SongId { get; set; }
