+++ src/Laul.Application/Services/Playlists/Commands/CreatePlaylist/CreatePlaylistCommand.cs
@@ -2,7 +2,7 @@
 
 namespace Laul.Application.Services.Playlists.Commands.CreatePlaylist
 {
-    public class CreatePlaylistCommand : IRequest<int>
+    public class CreatePlaylistCommand : IRequest<ulong>
     {
         public string Title { get; set; }
         public string Description { get; set; }
